Cache tag-dispatch vocabulary slicing (#28)
## Summary

- Cache the vocabulary slicing used by structural-tag dispatch.
- Reuse the second-slicing bitset (tokens that contain no trigger or
exclude string from the second character on) across compiles that share
the same trigger/exclude set. The per-state first-character masks are
computed as before.
- Preserve tag dispatch acceptance behavior.

## Review scope

This is a self-contained slice of mlc-ai/xgrammar#795 and is based
directly on main.

## Validation

- Relevant structural-tag and matcher tests pass in the isolated branch.
- Modal CPU screen measured 1.455x end-to-end speedup on structural
small.
- All compared outcomes matched main.

+++ cpp/support/utils.h
@@ -107,7 +107,7 @@ struct hash<std::tuple<Args...>> {
 template <typename T>
 struct hash<std::vector<T>> {
   size_t operator()(const std::vector<T>& vec) const {
-    uint32_t seed = 0;
+    uint64_t seed = 0;
     for (const auto& item : vec) {
       xgrammar::HashCombineBinary(seed, std::hash<T>{}(item));
     }
